nvme: narrow ns md_size to uint16_t
The metadata size cached in struct spdk_nvme_ns is sourced from
nsdata->lbaf[format_index].ms (see nvme_ns_set_identify_data() in
lib/nvme/nvme_ns.c). The NVMe spec defines the LBA Format metadata
size field as a 16-bit value, modeled in include/spdk/nvme_spec.h as
`uint32_t ms : 16`, so the value can never exceed 65535. Caching it
as a full uint32_t doubles the storage required for no benefit and
adds two padding bytes downstream.

This patch only changes the storage type. Field order and the
remaining holes are addressed in the next patch in the series.

Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/28775
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>

+++ lib/nvme/nvme_internal.h
@@ -587,7 +587,7 @@ struct spdk_nvme_ns {
 	 */
 	uint32_t			extended_lba_size;
 
-	uint32_t			md_size;
+	uint16_t			md_size;
 	uint8_t				pi_type			: 3;	/* \ref spdk_nvme_pi_type */
 	uint8_t				pi_format		: 2;	/* \ref spdk_nvme_pi_format */
 	uint32_t			sectors_per_max_io;
